+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 08-11-20
+### Changed
+- no-std compatibility
+- export scalar as `BlsScalar`
+
 ## [0.1.5] - 29-10-20
 ### Changed
 - Deriva Canon traits for Scalar